# swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "API для пірсингу",
      "description": "API для реєстрації користувачів, управління доступними днями та годинами для запису на послуги пірсингу.",
      "version": "1.0.0"
    },
    "servers": [
      {
        "url": "http://localhost:3000"
      }
    ],
    "paths": {
      "/register": {
        "post": {
          "summary": "Реєстрація нового користувача/адміністратора",
          "operationId": "register",
          "tags": ["Auth"],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "username": {
                      "type": "string",
                      "minLength": 5
                    },
                    "password": {
                      "type": "string",
                      "minLength": 8
                    },
                    "role": {
                      "type": "string",
                      "enum": ["user", "admin"],
                      "default": "user"
                    }
                  },
                  "required": ["username", "password"]
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Користувача зареєстровано"
            },
            "400": {
              "description": "Невірний формат запиту"
            },
            "500": {
              "description": "Виникла помилка на сервері"
            }
          }
        }
      },
      "/login": {
        "post": {
          "summary": "Логін користувача/адміністратора",
          "operationId": "login",
          "tags": ["Auth"],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "username": {
                      "type": "string"
                    },
                    "password": {
                      "type": "string"
                    }
                  },
                  "required": ["username", "password"]
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Успішний логін",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "token": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            },
            "401": {
              "description": "Неправильне ім'я користувача або пароль"
            },
            "500": {
              "description": "Виникла помилка на сервері"
            }
          }
        }
      },
      "/logout": {
        "post": {
          "summary": "Логаут користувача/адміністратора",
          "operationId": "logout",
          "tags": ["Auth"],
          "responses": {
            "200": {
              "description": "Успішний вихід з системи"
            },
            "500": {
              "description": "Виникла помилка на сервері"
            }
          }
        }
      },
      "/admin/free_days": {
        "post": {
          "summary": "Додавання доступних днів для запису (тільки для адміністраторів)",
          "operationId": "addFreeDays",
          "tags": ["Admin"],
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "date": {
                      "type": "string",
                      "format": "date"
                    },
                    "is_available": {
                      "type": "boolean"
                    }
                  },
                  "required": ["date", "is_available"]
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "День додано"
            },
            "400": {
              "description": "Не заповнені необхідні поля"
            },
            "403": {
              "description": "Недостатньо прав для виконання"
            },
            "500": {
              "description": "Виникла помилка на сервері"
            }
          }
        }
      },
      "/admin/free_hours": {
        "post": {
          "summary": "Додавання доступних годин для запису (тільки для адміністраторів)",
          "operationId": "addFreeHours",
          "tags": ["Admin"],
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "date": {
                      "type": "string",
                      "format": "date"
                    },
                    "time": {
                      "type": "string",
                      "format": "time"
                    },
                    "is_available": {
                      "type": "boolean"
                    }
                  },
                  "required": ["date", "time", "is_available"]
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Година додана"
            },
            "400": {
              "description": "Не заповнені необхідні поля"
            },
            "404": {
              "description": "День не знайдений"
            },
            "403": {
              "description": "Недостатньо прав для виконання"
            },
            "500": {
              "description": "Виникла помилка на сервері"
            }
          }
        }
      },
      "/free_days": {
        "get": {
          "summary": "Перегляд доступних днів для запису",
          "operationId": "getFreeDays",
          "tags": ["User"],
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "responses": {
            "200": {
              "description": "Перелік доступних днів",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "date": {
                          "type": "string",
                          "format": "date"
                        },
                        "is_available": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                }
              }
            },
            "404": {
              "description": "Немає доступних днів"
            },
            "500": {
              "description": "Виникла помилка на сервері"
            }
          }
        }
      },
      "/free_hours": {
        "get": {
          "summary": "Перегляд доступних годин для запису",
          "operationId": "getFreeHours",
          "tags": ["User"],
          "responses": {
            "200": {
              "description": "Перелік доступних годин",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "free_day_id": {
                          "type": "integer"
                        },
                        "time": {
                          "type": "string",
                          "format": "time"
                        },
                        "is_available": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                }
              }
            },
            "500": {
              "description": "Виникла помилка на сервері"
            }
          }
        }
      }
    },
    "components": {
      "securitySchemes": {
        "bearerAuth": {
          "type": "http",
          "scheme": "bearer",
          "bearerFormat": "JWT"
        }
      }
    }
  }
  